backend/erp_api/requirements.txt
# ML-Enabled ERP Inventory Intelligence Platform - Backend API
# Python 3.10+

# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Config & validation
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# ML-ready export & utilities
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
joblib==1.3.2

# Dev / testing
httpx>=0.26.0
pytest>=7.4.0,<9
pytest-asyncio>=0.23.0
